requirements.txt
alembic==1.13.3
amqp==5.2.0
billiard==4.2.1
blinker==1.8.2
celery==5.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
email-validator==2.1.1
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==3.1.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
MarkupSafe==3.0.1
packaging==24.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==4.6.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.20
stripe==11.1.0
supabase==2.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.7
WTForms==3.0.1
